refactor(web): migrate SoftwarePatternsSelection to new form conventions
- Move to software/patterns-selection/ subdirectory
- Rename SoftwarePatternsSelection.tsx to Form.tsx
- Create fields.ts with form options
- Rename softwarePatternsFormOptions to defaultOptions
- Move test JSON files to patterns-selection/
- Update imports in routes and tests
