+++ web/src/components/software/patterns-selection/patterns.test.json
