+++ web/src/components/software/patterns-selection/proposal.test.json
